Removes Customers page
Small dashboard fixes

Moves dashboard components into a dashboards folder
