+++ frontend/src/src/App.css
@@ -23,4 +23,5 @@ body {
 
 h1 {
   color: #004299;
+  margin-top: 2em;
 }
