--- frontend/src/src/components/CustomersPage.js
@@ -1,18 +1,0 @@
-import React from "react";
-import Navbar from "./Navbar";
-import "./componentStyles/CustomersPage.css";
-
-function CustomersPage() {
-  /*
-    Main page for /customers
-  */
-
-  return (
-    <div>
-      <Navbar />
-      <h1>Customer page</h1>
-    </div>
-  );
-}
-
-export default CustomersPage;
