--- frontend/src/src/components/componentStyles/AddCell.css
@@ -1,56 +1,0 @@
-.flex_container {
-  display: flex;
-  margin-top: 4em;
-  margin: 0 2em 2em 2em;
-}
-
-.options-wrapper {
-  display: flex;
-  flex-direction: column;
-  background-color: white;
-  min-width: 15em;
-  width: 30%;
-  text-align: center;
-}
-
-.graph_container {
-  min-width: 70em;
-  width: 100%;
-}
-
-.graph {
-  margin: 0 2em 2em 0;
-}
-
-.outer_flex_container {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-}
-
-.add_cell,
-.discard_cell {
-  margin: 6em 2em 1em 0.5em;
-}
-
-.add_cell_buttons {
-  display: flex;
-
-  align-self: flex-end;
-}
-
-.options-form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  flex-grow: 1;
-}
-
-#title, #primaryAxis, #secondaryAxis {
-  width: 80%;
-}
-
-.metadata_button {
-  margin-bottom: 2em;
-  justify-self: flex-end;
-}
