--- frontend/src/src/components/componentStyles/CustomersPage.css
@@ -1,3 +1,0 @@
-h1 {
-  margin-top: 2em;
-}
