--- frontend/src/src/components/componentStyles/LineGraph.css
@@ -1,3 +1,0 @@
-.chart_wrapper {
-    width: 100%;
-}
