+++ frontend/src/src/components/componentStyles/dashboards/DashboardCellCard.css
