+++ frontend/src/src/components/componentStyles/dashboards/LineGraph.css
@@ -1,0 +1,3 @@
+.chart_wrapper {
+    width: 100%;
+}
